refactor: various code optimizations & fixes
